chore: use uint16 for output index
We were using uint8 for output index which limits the tx output to
around 256. For DEX batch settlement this seems to be small. This commit
extends the field to uint16 and cap the output index to limit of
TX_OFFSET (10000).

- issue: https://github.com/omisego/plasma-contracts/issues/168

+++ plasma_framework/contracts/mocks/utils/UtxoPosLibWrapper.sol
@@ -15,7 +15,7 @@ contract UtxoPosLibWrapper {
         return UtxoPosLib.UtxoPos(_utxoPos).txIndex();
     }
 
-    function outputIndex(uint256 _utxoPos) public pure returns (uint8) {
+    function outputIndex(uint256 _utxoPos) public pure returns (uint16) {
         return UtxoPosLib.UtxoPos(_utxoPos).outputIndex();
     }
